📝 Group관련 Advice 및 Exception 분리
 Group관련 CRUD가 전부 다 끝나면 하나의 Advice, 상속을 통해 코드의 재사용성을 높힌 Exceptions로 리펙토링 예정
