+++ src/main/java/com/grepp/spring/infra/error/exceptions/group/GroupNotFoundException.java
@@ -1,18 +1,18 @@
-package com.grepp.spring.infra.error.exceptions;
+package com.grepp.spring.infra.error.exceptions.group;
 
 import com.grepp.spring.infra.response.GroupErrorCode;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-public class NotGroupUserException extends RuntimeException {
+public class GroupNotFoundException extends RuntimeException {
 
     private final GroupErrorCode code;
 
-    public NotGroupUserException(GroupErrorCode code) {
+    public GroupNotFoundException(GroupErrorCode code) {
         this.code = code;
     }
 
-    public NotGroupUserException(GroupErrorCode code, Exception e) {
+    public GroupNotFoundException(GroupErrorCode code, Exception e) {
         this.code = code;
         log.error(e.getMessage(), e);
     }
@@ -20,4 +20,5 @@ public class NotGroupUserException extends RuntimeException {
     public GroupErrorCode code() {
         return code;
     }
+
 }
